rename modelGenerators folder to modelTransfomers
